checkpoint: resolve stash-pop conflicts + concurrent WIP
Resolved stash-pop conflicts across 8 source files (Tail, VocalPlayer,
TokenRedemptionPanel, UIStats, BandDirector, StorePanel, Mesh, Part) —
unblocks build. Removed duplicate `bool replace;` in OutfitConfig MatSwap.
Folds in pending edits from parallel sweep agents across band3/ and system/.

Build green: 61.15% matched, 562/562, batch_check synced.

+++ src/band3/game/FadePanel.cpp
@@ -61,7 +61,7 @@ void FadePanel::Poll() {
 
 void FadePanel::Draw() {
     UIPanel::Draw();
-    static Hmx::Rect screen_rect(0, 0, TheRnd->mWidth, TheRnd->mHeight);
+    static Hmx::Rect screen_rect(0.0f, 0.0f, (float)TheRnd->mWidth, (float)TheRnd->mHeight);
     if (unk_0x38) {
         mColor.alpha = unk_0x84;
         TheRnd->DrawRect(screen_rect, mColor, NULL, NULL, NULL);
